% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_tr.R
\name{gene_tr}
\alias{gene_tr}
\title{Gene ID transform}
\usage{
gene_tr(type, ID)
}
\arguments{
\item{type}{ENTREZID or SYMBOLID}

\item{ID}{Set of genes named ENTREZID or SYMBOLID}
}
\value{
ENTREZID or SYMBOLID
}
\description{
Gene ID transform
}
\keyword{Gene}
\keyword{ID}
\keyword{transform}
